addr fb: move func from s3 to template
